feat(ui-plot): f64 domain and view-range sampling

+++ apps/fret-examples/src/area_demo.rs
@@ -35,7 +35,7 @@ impl AreaDemoDriver {
         let mut series1: Vec<DataPoint> = Vec::with_capacity(n);
 
         for i in 0..n {
-            let t = i as f32 / (n - 1) as f32;
+            let t = i as f64 / (n - 1) as f64;
             let x = t * 10.0;
 
             series0.push(DataPoint {
