+++ apps/fret-examples/src/plot_demo.rs
@@ -34,7 +34,7 @@ impl PlotDemoDriver {
         let mut series1: Vec<DataPoint> = Vec::with_capacity(n);
         let mut series2: Vec<DataPoint> = Vec::with_capacity(n);
 
-        let push = |series: &mut Vec<DataPoint>, x: f32, y: f32| {
+        let push = |series: &mut Vec<DataPoint>, x: f64, y: f64| {
             if !x.is_finite() || !y.is_finite() {
                 return;
             }
@@ -42,7 +42,7 @@ impl PlotDemoDriver {
         };
 
         for i in 0..n {
-            let t = i as f32 / (n - 1) as f32;
+            let t = i as f64 / (n - 1) as f64;
             let x = t * 10.0;
             push(
                 &mut series0,
